a little bit of research on 3.
I added a small explanation to the concepts from point 3. All of them
are by default done by StandardAnalyzer. Lucene gives the possibitity to
define our own analyzer, but dealing with natural language processing in
2 weeks maybe is too difficult. As an advance analysis we could add
synonyms. I done a little research on this and there is a package for
wordnet that maybe we can use. But once again this should be added only
after we have a prototype that works ok. Adding synonyms to the analyzer
is the extra-mile in my opinion so first we should concentrate only on
making an application that is good enough to be presented as a final
product